Add files via upload
Safe drinking water is essential to a healthy life. It is a fundamental human right. In order to predict weather the water quality is good enough for consumption, I used certain ML algorithms and models which comprised of Linear Regression, Decision Tree, KNN, SVM , Random Forest and AdaBoost. At the end of the project I got to know which model performed better in terms of accuracy by determining whether the quality of water is good or not.

I also used certain libraries like Panda's, Numpy, Matplotlib, Seaborn, Sk-learn and with the use of these I imported training and test data and I even plotted certain plots as well as the heatmaps for the easy visualization of the data.